LayoutTests/platform/wpe/storage/domstorage/localstorage/delete-removal-expected.txt

CONSOLE MESSAGE: QuotaExceededError: The quota has been exceeded.
This test makes sure that using the syntax `delete localStorage.keyName` works as an alias to `localStorage.removeItem(keyName).`
foo (before anything) is: undefined
